chore: refactor and intro run-metadata
