+++ tests/unit/adapters/targets/d2/test_d2_test_relationship.py
@@ -239,7 +239,7 @@ class TestD2TestRelationship:
             return_value=tables,
         ) as mock_get_tables:
             with mock.patch(
-                "dbterd.adapters.algos.test_relationship.get_relationships",
+                "dbterd.adapters.algos.base.get_relationships",
                 return_value=relationships,
             ) as mock_get_relationships:
                 mermaid = engine.parse(
