+++ tests/unit/adapters/targets/dbml/test_dbml_test_relationship.py
@@ -334,7 +334,7 @@ class TestDbmlTestRelationship:
             return_value=tables,
         ) as mock_get_tables:
             with mock.patch(
-                "dbterd.adapters.algos.test_relationship.get_relationships",
+                "dbterd.adapters.algos.base.get_relationships",
                 return_value=relationships,
             ) as mock_get_relationships:
                 dbml = engine.parse(
