+++ tests/unit/adapters/targets/mermaid/test_mermaid_test_relationship.py
@@ -231,7 +231,7 @@ class TestMermaidTestRelationship:
             return_value=tables,
         ) as mock_get_tables:
             with mock.patch(
-                "dbterd.adapters.algos.test_relationship.get_relationships",
+                "dbterd.adapters.algos.base.get_relationships",
                 return_value=relationships,
             ) as mock_get_relationships:
                 mermaid = engine.parse(
