+++ tests/unit/adapters/targets/plantuml/test_plantuml_test_relationship.py
@@ -237,7 +237,7 @@ class TestPlantUMLTestRelationship:
             return_value=tables,
         ) as mock_get_tables:
             with mock.patch(
-                "dbterd.adapters.algos.test_relationship.get_relationships",
+                "dbterd.adapters.algos.base.get_relationships",
                 return_value=relationships,
             ) as mock_get_relationships:
                 plantuml = engine.parse(
